requirements/development.txt
-r base.txt
django-debug-toolbar==1.11
django-extensions==2.1.6

